feat: require login for dashboard
