+++ frontend_updated_for_api/lib/auth-guard.tsx
@@ -1,0 +1,21 @@
+"use client"
+
+import { ReactNode, useEffect } from "react"
+import { useRouter } from "next/navigation"
+import { useAuth } from "@/lib/auth-context"
+
+export default function AuthGuard({ children }: { children: ReactNode }) {
+  const { isAuthenticated, loading } = useAuth()
+  const router = useRouter()
+
+  useEffect(() => {
+    if (!loading && !isAuthenticated) {
+      router.replace("/login")
+    }
+  }, [loading, isAuthenticated, router])
+
+  if (loading) return null
+  if (!isAuthenticated) return null
+  return <>{children}</>
+}
+
